[fix] 게시물에 접근하는 권한 처리 수정

+++ board_api/src/boards/boards.service.ts
@@ -6,6 +6,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { BoardRepository } from './board.repository';
 import { Board } from './board.entity';
 import { User } from 'src/auth/user.entity';
+import { GetUser } from 'src/auth/get-user.decorator';
 
 @Injectable()
 export class BoardsService {
@@ -19,15 +20,18 @@ export class BoardsService {
     ): Promise<Board[]> {
         const query = this.boardRepository.createQueryBuilder('board');
 
-        query.where('board.userId = :userid', { userId: user.id });
+        query.where('board.userId = :userId', { userId: user.id });
 
         const boards = await query.getMany();
 
         return boards;
     }
 
-    async getBoardById(id: number): Promise<Board> {
-        const found = await this.boardRepository.findOne(id);
+    async getBoardById(
+        id: number,
+        user: User
+    ): Promise<Board> {
+        const found = await this.boardRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Can't find Board with id ${id}`);
@@ -42,15 +46,22 @@ export class BoardsService {
         return this.boardRepository.createBoard(createBoardDto, user);
     }
 
-    async deleteBoard(id: number, user: User): Promise<void> {
-        const result = await this.boardRepository.delete({id, user});
+    async deleteBoard(
+        id: number,
+        user: User
+    ): Promise<void> {
+        const result = await this.boardRepository.delete({id, userId: user.id});
         if (result.affected === 0) {
             throw new NotFoundException(`Can't find Board with id ${id}`)
         }
     }
 
-    async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-        const board = await this.getBoardById(id);
+    async updateBoardStatus(
+        id: number,
+        status: BoardStatus,
+        user: User
+    ): Promise<Board> {
+        const board = await this.getBoardById(id, user);
 
         board.status = status;
         await this.boardRepository.save(board);
